INTERNAL-497 - Resolve CR comments
