+++ app/design/frontend/Satoshi/Hyva/Magento_LayeredNavigation/templates/layer/filter.phtml
@@ -37,7 +37,7 @@ $catalogHelper = $this->helper(Data::class);
                                 aria-label="<?= $escaper->escapeHtmlAttr(__('%1 filter, %2 available products', strip_tags((string)$filterItem->getLabel()), $filterItem->getCount())) ?>"
                             <?php endif; ?>
                         />
-                        <?php if ($filterItem->getName() === 'Color' && $filterItem->getSwatchValue()): ?>
+                        <?php if ($filterItem->getSwatchValue()): ?>
                             <span
                                 class="w-6 h-6 mr-5 p-1 rounded border border-gray-300"
                                 style="background-color: <?= $escaper->escapeHtml($filterItem->getSwatchValue()) ?>;"
